Rename multiCoreHandler to optimizer.js
